refactor(dev): extract DevToolbar row components to DevToolbarRows
Behavior-preserving move (DevToolbar.tsx 1465 -> 1280): the prop-driven
row/toggle components (PlanToggle, PlanToggleInner, OrphanOverrides,
FlagRow) plus the useSafeQueryClient helper move to a sibling
DevToolbarRows.tsx (dev-only, same feature dir). These are self-contained
(only shared dep useSafeQueryClient moved with them; verified none of the
main body uses it). Now-unused imports (Switch, queryKeys,
useBillingStatusQuery, useQueryClientBase, QueryClient type) dropped from
DevToolbar.tsx and moved to the new file. Bodies verbatim.
